backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.0",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "fastapi>=0.111.1",
    "uvicorn>=0.30.1",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
]
